+++ src/main/kotlin/app/revanced/patches/youtube/misc/integrations/fingerprints/InitFingerprint.kt
@@ -3,5 +3,5 @@ package app.revanced.patches.youtube.misc.integrations.fingerprints
 import app.revanced.patches.shared.integrations.patch.AbstractIntegrationsPatch.IntegrationsFingerprint
 
 object InitFingerprint : IntegrationsFingerprint(
-    strings = listOf("Application creation"),
+    strings = listOf("Application creation", "Application.onCreate"),
 )