+++ src/main/kotlin/app/revanced/patches/youtube/misc/litho/filter/patch/LithoFilterPatch.kt
@@ -1,10 +1,12 @@
 package app.revanced.patches.youtube.misc.litho.filter.patch
 
+import app.revanced.extensions.toErrorResult
 import app.revanced.patcher.annotation.Description
 import app.revanced.patcher.annotation.Version
 import app.revanced.patcher.data.BytecodeContext
 import app.revanced.patcher.extensions.addInstructions
 import app.revanced.patcher.extensions.instruction
+import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
 import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint.Companion.resolve
 import app.revanced.patcher.patch.BytecodePatch
 import app.revanced.patcher.patch.PatchResult
@@ -16,12 +18,10 @@ import app.revanced.patches.youtube.misc.integrations.patch.IntegrationsPatch
 import app.revanced.patches.youtube.misc.litho.filter.annotation.LithoFilterCompatibility
 import app.revanced.patches.youtube.misc.litho.filter.fingerprints.ComponentContextParserFingerprint
 import app.revanced.patches.youtube.misc.litho.filter.fingerprints.EmptyComponentBuilderFingerprint
+import app.revanced.patches.youtube.misc.litho.filter.fingerprints.ReadComponentIdentifierFingerprint
 import org.jf.dexlib2.iface.instruction.Instruction
 import org.jf.dexlib2.iface.instruction.OneRegisterInstruction
 import org.jf.dexlib2.iface.instruction.ReferenceInstruction
-import org.jf.dexlib2.iface.instruction.TwoRegisterInstruction
-import org.jf.dexlib2.iface.reference.FieldReference
-import org.jf.dexlib2.iface.reference.MethodReference
 
 @DependsOn([IntegrationsPatch::class])
 @Description("Hooks the method which parses the bytes into a ComponentContext to filter components.")
@@ -31,33 +31,40 @@ class LithoFilterPatch : BytecodePatch(
     listOf(ComponentContextParserFingerprint)
 ) {
     override fun execute(context: BytecodeContext): PatchResult {
-        ComponentContextParserFingerprint.result?.let { result ->
-            val builderMethodIndex = EmptyComponentBuilderFingerprint
-                .also { it.resolve(context, result.mutableMethod, result.mutableClass) }
-                .let { it.result!!.scanResult.patternScanResult!!.startIndex }
-
+        ComponentContextParserFingerprint.result?.also {
+            arrayOf(EmptyComponentBuilderFingerprint, ReadComponentIdentifierFingerprint).forEach { fingerprint ->
+                if (!fingerprint.resolve(context, it.mutableMethod, it.mutableClass))
+                    return fingerprint.toErrorResult()
+            }
+        }?.let { result ->
+            val builderMethodIndex = EmptyComponentBuilderFingerprint.patternScanEndIndex
             val emptyComponentFieldIndex = builderMethodIndex + 2
 
-            with(result.mutableMethod) {
+            result.mutableMethod.apply {
                 val insertHookIndex = result.scanResult.patternScanResult!!.endIndex
-                val clobberedRegister = (instruction(insertHookIndex - 3) as OneRegisterInstruction).registerA
+                val builderMethodDescriptor = instruction(builderMethodIndex).descriptor
+                val emptyComponentFieldDescriptor = instruction(emptyComponentFieldIndex).descriptor
+                // Register is overwritten right after it is used for this patch, therefore free to clobber.
+                val clobberedRegister = instruction(insertHookIndex).oneRegister
 
-                val builderMethodDescriptor = instruction(builderMethodIndex).toDescriptor()
-                val emptyComponentFieldDescriptor = instruction(emptyComponentFieldIndex).toDescriptor()
+                @Suppress("UnnecessaryVariable")
+                // The register, this patch clobbers, is previously used for the StringBuilder,
+                // later on a new StringBuilder is instantiated on it.
+                val stringBuilderRegister = clobberedRegister
 
-                val stringBuilderRegister = (instruction(insertHookIndex - 1) as TwoRegisterInstruction).registerA
+                val identifierRegister = instruction(ReadComponentIdentifierFingerprint.patternScanEndIndex).oneRegister
 
                 addInstructions(
                     insertHookIndex, // right after setting the component.pathBuilder field,
                     """
-                        invoke-static {v$stringBuilderRegister, v0}, Lapp/revanced/integrations/patches/LithoFilterPatch;->filter(Ljava/lang/StringBuilder;Ljava/lang/String;)Z
+                        invoke-static {v$stringBuilderRegister, v$identifierRegister}, Lapp/revanced/integrations/patches/LithoFilterPatch;->filter(Ljava/lang/StringBuilder;Ljava/lang/String;)Z
                         move-result v$clobberedRegister
                         if-eqz v$clobberedRegister, :not_an_ad
-                        move-object/from16 v0, p1
-                        invoke-static {v0}, $builderMethodDescriptor
-                        move-result-object v0
-                        iget-object v0, v0, $emptyComponentFieldDescriptor
-                        return-object v0
+                        move-object/from16 v$clobberedRegister, p1
+                        invoke-static {v$clobberedRegister}, $builderMethodDescriptor
+                        move-result-object v$clobberedRegister
+                        iget-object v$clobberedRegister, v$clobberedRegister, $emptyComponentFieldDescriptor
+                        return-object v$clobberedRegister
                     """,
                     listOf(ExternalLabel("not_an_ad", instruction(insertHookIndex)))
                 )
@@ -68,14 +75,14 @@ class LithoFilterPatch : BytecodePatch(
     }
 
     private companion object {
-        fun Instruction.toDescriptor() = when (val reference = (this as? ReferenceInstruction)?.reference) {
-            is MethodReference -> "${reference.definingClass}->${reference.name}(${
-                reference.parameterTypes.joinToString(
-                    ""
-                ) { it }
-            })${reference.returnType}"
-            is FieldReference -> "${reference.definingClass}->${reference.name}:${reference.type}"
-            else -> throw PatchResultError("Unsupported reference type")
-        }
+        val MethodFingerprint.patternScanEndIndex
+            get() = result!!.scanResult.patternScanResult!!.endIndex
+
+        val Instruction.descriptor
+            get() = (this as ReferenceInstruction).reference.toString()
+
+        val Instruction.oneRegister
+            get() = (this as OneRegisterInstruction).registerA
+
     }
 }