+++ src/main/kotlin/app/revanced/patches/youtube/misc/video/speed/remember/fingerprint/InitializePlaybackSpeedValuesFingerprint.kt
@@ -1,7 +1,8 @@
-package app.revanced.patches.youtube.misc.video.speed.remember.fingerprint
+package app.revanced.patches.youtube.misc.video.speed.current.fingerprint
 
 import app.revanced.patcher.fingerprint.method.impl.MethodFingerprint
 
 object InitializePlaybackSpeedValuesFingerprint : MethodFingerprint(
-    parameters = listOf("[L", "I")
+    parameters = listOf("[L", "I"),
+    strings = listOf("menu_item_playback_speed")
 )